Restore original method hierarchy
