+++ app/models/concerns/with_expectations.rb
@@ -18,14 +18,6 @@ module WithExpectations
     self[:expectations] = expectations.map(&:stringify_keys)
   end
 
-  def own_expectations
-    self[:expectations]
-  end
-
-  def own_custom_expectations
-    self[:custom_expectations]
-  end
-
   def ensure_expectations_format
     errors.add :own_expectations,
                :invalid_format unless own_expectations.to_a.all? { |it| Mulang::Expectation.valid? it }
