+++ app/models/exercise/challenge.rb
@@ -8,6 +8,14 @@ class Challenge < Exercise
     self.layout = self.class.default_layout
   end
 
+  def extra(*)
+    [guide.extra, super]
+      .compact
+      .join("\n")
+      .strip
+      .ensure_newline
+  end
+
   private
 
   def defaults
